inst/python/requirements.txt
# Python dependencies for the RabAnalyser feature extraction backend
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-image>=0.18.0
tifffile>=2021.1.0
matplotlib>=3.3.0
joblib>=1.0.0
python-igraph>=0.10.0
leidenalg>=0.9.0
umap-learn>=0.5.0
scikit-learn>=1.8.0
